drop MemberNameResolver for ContactsRepository.contactName
