+++ docs/plans/contact-list-prd.md
@@ -795,17 +795,17 @@ Ship-readiness: tapping a member from inside a chat opens the unified card; exis
 
 Stopgap for the window between "we invite a contact" and "their profile snapshot arrives." When an in-group render site would emit "Somebody" because the per-conversation profile is empty, fall back to the contacts-list display name first. Read-only; never modifies the contacts table; per-conversation profile always takes precedence once observed.
 
-- [ ] Add a `MemberNameResolver` (small helper, `Convos/` target — the `Profile` API in ConvosCore stays unchanged so its semantics don't drift across modules). Signature: `func resolvedDisplayName(for inboxId: String, profileName: String?) -> String`. Returns `profileName` verbatim if non-empty, otherwise the contact's `displayName` if non-empty, otherwise `"Somebody"`. Constructed with a `ContactsRepositoryProtocol` injected at init.
-- [ ] Inject the resolver into render sites that surface a remote member's name via `Profile.displayName`. Audit during implementation by grepping `Profile.displayName` callers inside `Convos/Conversation Detail/`. Initial known list:
+- [ ] Add a `contactName(for inboxId: String) -> String?` method to `ContactsRepositoryProtocol` (the `Profile` API in ConvosCore stays unchanged so its semantics don't drift across modules). Returns the contact's `displayName` if non-empty, otherwise `nil`. Non-throwing; swallows storage errors as `nil` since render-site callers cannot usefully handle a thrown error mid-paint. Composed with the per-conversation profile name and the "Somebody" sentinel by the existing `Conversation.computedDisplayName(memberNameOverride:)` / `ConversationMember.displayName(memberNameOverride:)` overloads in ConvosCore.
+- [ ] Inject the lookup into render sites that surface a remote member's name via `Profile.displayName`. Each surface constructs `messagingService.contactsRepository().contactName(for:)` once and threads the resulting `@Sendable (String) -> String?` closure through the SwiftUI `.memberNameOverride(_:)` environment modifier (defined in `Convos/Contacts/MemberNameOverride.swift`). Audit during implementation by grepping `Profile.displayName` callers inside `Convos/Conversation Detail/`. Initial known list:
   - System / update messages: join, leave, added-by, removed-by, role change, name change. These typically share a renderer; updating it once covers all the variants.
   - The conversation members list (`ConversationMembersListView` → `MemberRow.displayName`).
   - Mention strings rendered in message bodies (if any).
   - Typing-indicator labels that include a member name (verify these route through `Profile.displayName` first).
-- [ ] Skip the resolver for any render site that's known to be "self-only" (the local user always has their own profile populated). Don't add overhead to rendering of self.
+- [ ] Skip the lookup for any render site that's known to be "self-only" (the local user always has their own profile populated). Don't add overhead to rendering of self.
 - [ ] Don't change `Profile.displayName` itself. The existing "Somebody" behavior remains for callers that haven't opted in. This keeps the fallback opt-in per render site rather than a global semantics shift, and prevents log lines / analytics / unit tests that read `displayName` from suddenly seeing contact names.
 - [ ] Out of scope for V1: NSE / push-notification copy and any rendering outside the conversation view that doesn't have access to a `ContactsRepositoryProtocol`. Those targets continue rendering "Somebody"; a follow-up can plumb the repo through.
 - [ ] Tests:
-  - `MemberNameResolver` unit tests: per-conversation name takes precedence; fallback to contact when name is empty; "Somebody" when neither is available; no contact write occurs during a lookup.
+  - `ContactsRepository.contactName(for:)` unit tests: returns the stored display name when the inbox is a known contact, returns `nil` for non-contacts and for contacts with an empty `displayName`, swallows storage errors as `nil`; no contact write occurs during a lookup. Composition with the per-conversation profile name and the "Somebody" sentinel is already covered by the `MemberNameOverrideTests` suite in ConvosCore.
   - Snapshot- or string-equality test on the system-message renderer for "Somebody joined" → "Alice joined" transformation when Alice is a contact.
 
 Ship-readiness: a member who joins a group I share with them shows by their contact name in system messages and the members list, even before their per-conversation profile snapshot lands; non-contacts continue to render as "Somebody"; no behavior change for the local user's own avatar / name.
